Move request store to common
