+++ src/vs/platform/terminal/common/requestStore.ts
